requirements.txt
streamlit
spacy
scikit-learn
pdfplumber
python-docx
pandas
rapidfuzz
matplotlib
joblib